requirements.txt
__future__
kaleido>=1.1.0,<2.0.0
matplotlib>=3.10.1,<4.0.0
numpy>=2.2.1,<3.0.0
Pillow>=11.1.0,<12.0.0
plotly>=6.3.0,<7.0.0
PyYAML>=6.0.2,<7.0.0
